Fix warnings about alteration of constant value
git-svn-id: https://svn.osgeo.org/gdal/branches/2.2@38009 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/hfa/hfadataset.cpp
@@ -2182,8 +2182,10 @@ void HFARasterBand::ReadHistogramMetadata()
         for( int i = 0; i < nNumBins; i++ )
         {
             const double dfNumber = padfWorkBuf[i];
-            if( dfNumber >= std::numeric_limits<GUIntBig>::max() ||
-                dfNumber < std::numeric_limits<GUIntBig>::min() ||
+            if( dfNumber >= static_cast<double>(
+                                std::numeric_limits<GUIntBig>::max()) ||
+                dfNumber < static_cast<double>(
+                                std::numeric_limits<GUIntBig>::min()) ||
                 CPLIsNan(dfNumber) )
             {
                 CPLError(CE_Failure, CPLE_FileIO, "Out of range hist vals.");
